scripts/requirements.txt
torch
torchvision
timm
albumentations
scikit-learn
matplotlib
pillow
pandas
numpy
jupyterlab
ipykernel



flask



